+++ R/argparse.R
@@ -68,6 +68,26 @@ ArgumentParser <- function(..., python_cmd = NULL) { # nolint
         "except ImportError:",
         "    import simplejson as json",
         "",
+        "def as_logical(s):",
+        "    if isinstance(s, bool):",
+        "        return bool",
+        "    elif s in ('T', 'TRUE', 'True', 'true'):",
+        "        return True",
+        "    elif s in ('F', 'FALSE', 'False', 'false'):",
+        "        return False",
+        "    else:",
+        "        return [None]", # {jsonlite} workaround
+        "",
+        "def as_logical_append(s):", # {jsonlite} workaround
+        "    if isinstance(s, bool):",
+        "        return bool",
+        "    elif s in ('T', 'TRUE', 'True', 'true'):",
+        "        return True",
+        "    elif s in ('F', 'FALSE', 'False', 'false'):",
+        "        return False",
+        "    else:",
+        "        return None",
+        "",
         sprintf("parser = argparse.ArgumentParser(%s)",
                 convert_..._to_arguments("ArgumentParser", ...)),
         "")
@@ -235,13 +255,12 @@ get_python_type <- function(type, proposed_arguments) {
             character = "str",
             double = "float",
             integer = "int",
-            logical = "bool",
+            logical = "as_logical",
             stop(paste(sprintf("type %s not supported,", type),
                     "supported types:",
                     "'logical', 'integer', 'double' or 'character'")))
-    # warn if type set to "logical" and action set to "store"
-    if (python_type == "bool" && any(grepl("action='store'", proposed_arguments)))
-        warning("You almost certainly want to use action='store_true' or action='store_false' instead")
+    if (python_type == "as_logical" && any(grepl("action='append'", proposed_arguments)))
+        python_type <- "as_logical_append" # {jsonlite} workaround
     sprintf("type=%s", python_type)
 }